Properly support startup notify when opening files or the bulk renamer.
When calling thunar from the command line, it now uses the
DESKTOP_STARTUP_ID environment variable and passes it to the functions
responsible to open/execute files and to open the bulk renamer.

An additional startup id paremeter is added to the functions

  thunar_application_bulk_rename()
  thunar_application_open_window()
  thunar_application_process_filenames()
  thunar_file_launch()
  thunar_show_renamer_dialog()
  thunar_dbus_client_launch_files()
  thunar_dbus_service_launch_files()

as well as to the following D-Bus methods:

  org.xfce.FileManager.Launch()
  org.xfce.Thunar.BulkRename()
  org.xfce.Thunar.LaunchFiles()

This was made possible by Nick who provided the entire patch.