Merge the migration-to-gio branch into master.
I'm 100% sure I can implement all ThunarVFS thumbnailers on top of
tumbler in time for the Xfce 4.8 release. Same about the volume monitor
backend for GIO to avoid a GVfs dependency. So why not merge the GIO
branch now?

Not much to say about this merge. It replaces ThunarVFS with GIO inside
Thunar and Thunarx, something I've been working on over the past few
months. It currently requires GVfs for volume monitoring and also
(optionally) depends on tumbler for thumbnail generation. Enjoy.

All .po files are going to be updated in the next commit.

Conflicts:
	ChangeLog
	INSTALL
	acinclude.m4
	configure.in.in
	plugins/thunar-wallpaper/twp-provider.c
	po-doc/ChangeLog
	po/ChangeLog
	po/LINGUAS
	thunar-vfs/thunar-vfs-1.pc.in
	thunar-vfs/thunar-vfs-thumb.c
	thunar/thunar-templates-action.c