Make emptying the trash work again.
With GIO, the size of the trash:/// directory no longer corresponds to
whether the trash is empty or not. Instead, we have to use the
G_FILE_ATTRIBUTE_TRASH_ITEM_COUNT attribute. This commit adds a new
method thunar_file_get_item_count() which is used the shortcuts pane and
the tree view to make the empty trash action sensitive or insensitive.

Also remove a few redundant g_file_info_has_attribute() checks.