Ignore cancel errors when processing the input filenames.
Cancelling e.g. the credentials dialog is something users do on purpose.
Thus we don't need to show them a cancel error dialog. There is a bug in
GVfs however which results in a permission denied error instead of a
cancel error when the user presses the cancel button in the credentials
dialog. It was filed against GVfs (#594798 in GNOME's bugzilla).