Initial import of the shared progress dialog.
This introduces a new view called ThunarProgressView which essentially
replaces the old ThunarProgressDialog. The new ThunarProgressDialog is a
container GtkWindow (I wonder about renaming it to ThunarProgressWindow)
for all the ThunarProgressViews associated with running file operations.

ThunarProgressDialog also creates a status icon that can be used to
toggle its visibility.

Things left on the TODO list:
- Check if we can lower the CPU usage of the dialog (X goes up to
  around 50% on my machine with the dialog visible), compare this to
  the situation before.
- Double-check if the use of gtk_window_present() is correct here.
- Either prepend (instead of append) new views or scroll down to them
  upon their creation.