Use -export-symbols-regex instead of XDT_FEATURE_VISIBILITY.
XDT_FEATURE_VISIBILITY() breaks plugins as it strips thunar_extension*
functions from the final shared library files. An alternative way to
XDT_FEATURE_VISIBILITY() and the old visibility configure code used in
Thunar is to define -export-symbols-regex (e.g. by setting it to
"^thunar_extension.*") in Makefile.am of each plugin.

This also removes the alias stuff in Thunarx, as we no longer need to
define which symbols to export or not via the alias header. All symbols
starting with _ are now automatically marked as private and are not
exported. Additionally, G_GNUC_INTERNAL can be used to hide symbols.

thunarx.symbols is preserved for ABI checks and still has to be updated
when new symbols are added.

Hope all this is ok. Double-checkers to the rescue!