Build manual documentation during make.
This will generate the manual during make if --enable-gen-doc
is used and only ship the generated html files in the dist
tarball. It will automatically pick up new translations (see
bug #6036) and print xml valid warnings which are useful for
translators to valid their translations.

All documentation is now also installed in @docdir@ (bug #5810).

Symlink to C images if the language does not ship its own
screenshots. This will reduce the tarball and installation
size a lot.

To keep the patch small, the existing translation directories are not
removed. Please do this after appying the patch:
  git rm -fr docs/manual/{C,da,es,eu,fr,gl,it,ja,nl,pl,ru,tr,zh_TW}
Optionally the po-doc/ChangeLog files should be deleted too.

Signed-off-by: Nick Schermer <nick@xfce.org>
Signed-off-by: Jannis Pohlmann <jannis@xfce.org>