Refilter the tree model when a hidden dir is entered/left (bug #6142).
What's new is that instead of only refiltering when entering a directory
that is within a hidden hierarchy, we now do the same when leaving a
hidden directory or directory with hidden ancestor. This makes sure
hidden directories are properly removed from the tree view side pane
when hidden directories are turned off, unless you're browsing them.