Create status icon when we're actually going to display it.
The show/hide in status icons emits a critical warning (in the
plug window) because of the quick hide we do after creating in
status icon. This crashes the status icon in the panel.
Although this is a problem in Gtk, it is also better to delay
creating the status icon until we actually show it.