Add CopyTo, CopyInto, MoveInto, LinkInto, UnlinkFiles D-Bus methods.
The groundwork for this was done by Daniel Morales. I refactored the
code a little bit, fixed a few bugs and added a working directory
parameter to all functions. The startup id parameter is unused at
the moment but this may change in the future.

The methods have been added to the org.xfce.FileManager D-Bus interface.