Delay setting the selection until the view has finished loading.
This fixes issues when calling thunar_component_set_selected_files() on
a view that has not finished loading the folder contents yet. Once the
folder has been loaded, the selection is always reset, so to work around
this, we need to delay setting the selected files until loading is done.