Auto-select subdirectories when walking through the history (bug #5538).
When entering a new directory, Thunar now checks if the next or previous
directory in the history is an immediate child of the new directory. If
one of these files is a child, it selects it automatically.

This makes keyboard navigation much easier. I'm sure the behaviour can
be confusing in some situations as well, but I hope it is an
improvement.

This partly fixes bug #5538 as well.