Implement a new D-Bus method RenameFile(uri, display, startup-id).
This method can be used by applications like xfdesktop in order to
prompt the user to choose a new name for a file. This way xfdesktop
doesn't have to implement the rename dialog and method itself anymore,
which results in a more consistent UI and only one place would have to
be changed if this UI needs to be revamped later.