Change how Thunar picks the default handler for a file (bug #6167).
It appears that g_file_query_default_handler() prefers URI scheme
handlers over content type handlers. This is inappropriate if we want to
open a file for which we already know the content type (as is the case
with most local files).

Using g_file_query_default_handler() as the only way to pick the default
handler for a file somewhat collides with exo-gio-module, which sets the
default handler for a few URI schemes, such as http:// and file://.

This was problematic only when launching a file from outside (e.g. by
running "Thunar <file>" or by calling the D-Bus Launch() method. In
internal situations, Thunar used thunar_file_list_get_applications()
which prioritizes the content type.

We now use g_app_info_get_default_for_content_type() first and only fall
back to g_file_query_default_handler() if the content type is unknown or
we don't have a default handler for it.