Fix the home folder being displayed as "Desktop" if home == desktop.
If the home and desktop folders are the same, then we want to show the
home folder as "username" rather than as "Desktop" in the shortcuts
pane. So we will no longer apply special label treatment to the desktop
folder (I guess GIO would internally figure out the localized desktop
display name based on xdg-user-dirs anyway...).