Properly try alternative names when creating links (bug #6867).
This was a regression compared to the Thunar code based on ThunarVFS.
When creating a symbolic link with a name for which a file already
exists, we need to try alternative names such as "link to foo", "another
link to foo" and "4th link to foo", just like we do when copying files.

This commit hopefully fixes this and also makes the "Make link(s)"
action in the "Edit" menu work again.