Only erase top-level items from trash (bug #7147).
Apparently, the GVfs trash backend only allows top-level directories in
the trash to be modified. Children can be deleted implicitely but e.g.
when emptying the trash, only the top-level items may be deleted.

To give an example: g_file_delete() fails for "trash:///test/foo.txt"
but succeeds for "trash:///test".

This problem is covered in the trash specification under
"Implementation notes".