Use a timeout/idle to request thumbnails in ThunarStandardView.
This is the second part of redesigning how thumbnails are requested in
Thunar. Each ThunarStandardView now has a ThunerThumbnailer object.
On scroll and resize events and whenever a new directory has finished
loading, ThunarStandardView now schedules a timeout or idle handler
(depending on the situation) to request thumbnails for all visible items
of the view.

This timeout/idle handler is rescheduled whenever the user starts
or continues scrolling or resizing. This avoids sending requests while
the user is scrolling and its important to not have D-Bus interfere with
the Thunar main loop. We now have a much more responsive thumbnailing
and scrolling experience!

Requests are also dropped by each view when a directory is closed/left.
This means that we no longer generate thumbnails the user is not
interested in any more.

This commit also uses the proper D-Bus method for cancelling thumbnail
requests (it's called "Dequeue", not "Unqueue").

The properties dialog is updated to match changes in the
ThunarThumbnailer API.