Add a ThunarThumbnailCache to ThunarApplication, implement Move().
With this tumblerd can just copy the thumbnail and adjust its meta data
when a file is renamed. We thereby avoid regenerating thumbnails for
files that are simply renamed but whose contents don't change at all.

We probably need some extra code to check whether the file type has now
changed or become incompatible with tumbler. In that case we will have
to drop the thumbnail. But how often does this really happen?