Speed up initial start by adding the Network item later (bug #7313).
There are other ways to achieve the same effect: load the icon and the
name of items in the side pane first and only resolve the real files
when users click on them. But this was easier. Also, I expect the side
pane to be redesigned soon and this fix makes it easier to apply the
change to the 4.8 branch.