Prevent looping in some renamers.
Some renamers trigger file changes when processing the
renamer name. Prevent this by checking if the disk content
really changed.

This is for example triggered by the thunar-media-tags-plugin
in the bulk renamer. For an unknown reason, each time a file
is read, the "changed" signal is triggered on the folder, resulting
in endless looping on the file, consuming 100% cpu.