Move away from using a hard-coded size for window icons (bug #8626).
The previous commit wasn't actually related to this bug. The problem
here was that the xfwm4 tab window now allows themes to change the size
of the displayed window icons but Thunar still hard-coded its own window
icons to 48px.

From this commit Thunar on uses thunar_file_get_icon_name(current_dir)
to get an icon name for the current directory and and then calls
gtk_window_set_icon_name() rather than asking ThunarIconFactory for a
48px GdkPixbuf that cannot be scaled up nicely. This is ok because we
only ever set icons for directories and for those ThunarIconFactory
never returns anything but a GdkPixbuf based on an icon name anyway. It
never returns a GdkPixbuf loaded from an absolute path or anything.