Rework the internals of the thumbnailer.
Use a single structure to follow a request inside the thumbnailer
instead of 3 hashtable. This makes the code easier to understand
and the hashtables were a bit overdue because there are never more
then n_windows running at the same time.

Also properly implement dequeue. Jobs were never properly aborted
and also the returned uris were emitted inside thunar, even if
the file was not visible anymore.

Don't emit file changes of jobs that are not known in the active
thumbnailer. Previously a file change was emitted n_windows-times
even if the other windows didn't even display the file.