Add ThunarDevice and ThunarDeviceMonitor.
This is an shell around GVolumeMonitor and GVolume/GMount and
possibly in the future GDrive. The idea is that the models
and views don't need to know what type they handle.

If a volume/drive/mount needs to be visible to the user, it is
added with device-added, if something changed, device-removed is
triggered and when the device need to be invisible or is removed
device-removed is called.

It also takes care of the mount/unmount/eject feature, independent of
the type.