Directly collect the full info namespace for ls dir.
Adapt thunar_io_scan_directory to optionally return
ThunarFile's instea of GFile and create ThunarFiles
later for some io jobs.

This trick is that we can directly use the enumerator
to scan all the directory info and use that for the
new ThunarFiles, this saves querying a file twice
resulting in much faster directory loading.