Store cairo surfaces on pixbufs.
Because we store pixbufs in the icon factory
and re-use them, its efficient to store the
generated cairo surfaces on the pixmaps and
re-use those as well.

For this the gdk_cairo_set_source_pixbuf has
been included in thunar to build the surface,
we then peek it if already generated for a
pixmap.