Update xfconf setting for thunar-wallpaper plugin
Xfdesktop uses a new xfconf format starting with version 4.11.
This patch updates the thunar-wallpaper plugin to set the image
in the new format while continuing to set the old format.

Comment from Jannis: This currently only works if "Apply to all
workspaces" is disabled in xfdesktop. We might have to set the
workspace0 wallpaper as well in order for the wallpaper to be
applied in this case.