Make single middle click open files and folders (bug #10587)
Regardless of the "open on single click" option, double middle
click is weird and many users will not know about it. What's more,
on many mice the middle mouse button is hard to press.

So do away with the double middle click and only allow open
with middle click for folders.