Allow <ctrl> + middle click in shortcut and treeview panes
Currently, only middle click is possible, without any modifiers.
This patch enables <ctrl> + middle click, enabling the user
to quickly switch (for that action only) the behaviour of
opening folders to either "open in new window" or "open in new
tab". This makes this behaviour equal to that of the
icons/list/details views.