Improve keyboard navigation for thunar-tree-view (bug #4519)
Left and right arrow (keypad too) expand/collapse, up and down arrow
work as usual, and the right pane follows the tree view selection.

There is currently a problem with the top-level Desktop item which
would result in opening the item in another subtree (the Desktop
usually resides in /home/user/Desktop). This way, opening the
item automatically would break tree view navigation and this is
bad for user experience. So to open this item, the user has to
press space, enter or keypad enter.