Sidepane treeview: Do not switch folder when clicking a row expander
With the changes introduced in commit f60acf89a1321be0faa37bfb0d91b1757085ffa4
"Allow <ctrl> + middle click in shortcut and treeview panes", clicking
the expander also selects the item and thus shows the content in the
standard view. This commit fixes this and restores the previous behaviour
for the expanders.