Fix updating file info for moved files (bug #11008)
When a file has been moved, it will trigger a G_FILE_MONITOR_EVENT_MOVED
in thunar-folder.c, with the target file saved in other_file. While we
reload the folder, that only fetches the info for the target file from
the file info cache, which still contains the information of the replaced
file. So issue a reload of the target file to update the file information
properly, and also a reload of the folder containing the target file if
it is not the same as the folder of the source file.