Improve drag & drop and selecting in the details view (bug #11605)
When using the left mouse button, clicking on the first column of an item
will select it or initiate dragging, but rubber banding is deactivated.
For the other columns, only rubber banding will be possible and dragging
be disabled.

When using the right mouse button, clicking on the first column of an
item will always select it and pop up the context menu of the item.
Clicking on another column of an unselected item will unselect all
items and show the menu of the folder, while clicking on another
column of a selected item will show the context menu for that item
(or those items).

An extensive explanation has been added to the thunar docs in the wiki:
http://docs.xfce.org/xfce/thunar/working-with-files-and-folders