Extend reload functions to allow a full reload
Usually a folder is reloaded automatically on file changes, and the user
shouldn't have to do it manually. However, when the user *wants* to do
it manually, everything is expected to be reloaded, even the files in
cache, because their info may have changed. At the moment, this does not
happen, and there can be various issues with cached file info, so extension
is a first step to allow the user to decide to reload everything in the
folder.