Let row changes accumulate a bit before restoring selection
With the changes introduced in 4d19f45feb198 "Restore selection on row
changes too (bug #9886)", slight flicker can be noticed when navigating
with forward/back actions. This is because row-changed events sometimes
cause a restore of the selection multiple times.

To solve this, this commit adds a short timeout for restoring the
selection in a similar way as for the status bar text updates.