Restore selection only when rows-reordered signal is received (bug #11855)
Commit 4d19f45feb198b "Restore selection on row changes too (bug #9886)"
forced the file selection to be restored on row changes, which has the
negative side effect that the view will also scroll to the selected
file(s) whenever such a change happens, e.g. when thumbnailing finishes.

The better fix is to listen to the rows-reordered signal which no view
listens to at the moment. This commit implements this.