Keep the scroll position when restoring the file selection
Instead of trying to keep the selected files in view when rows get
reordered, leave the scroll position unchanged. Otherwise, the user
might not be able to select files easily.

By setting fix lower/upper limits for the scrolled window alignments
we make sure that the view does not flicker because the position
would change twice.