Fix escaping and unescaping URI in path entry (bug #8910)
Escaping the URI in thunar_path_entry_changed is necessary for correctly
loading the corresponding ThunarFile for the GFile.

This solves bug #8910 where the problem stems from URI unescaping,
resulting in the unescaped text being NULL. This is handled by the
second part of the patch, which falls back to the unescaped text.

Without this commit, directories containing characters like "%", e.g
"test%folder", can be opened but

1) on remote locations the icon in the toolbar location selector will
   show an error indicating there is something wrong,
2) while on local filesystems only the parent path will be shown in
   the path entry.