Always select the proper toplevel item in the thunar tree view
When the user clicks on e.g. "Desktop" in the treeview, the treeview
will expand the tree and set the cursor on home/Desktop instead on the
proper toplevel item (assuming the Desktop folder resides in the user's
home directory of course). This is very bad for user experience.

To fix this, this commit makes the treeview

a) use the cursor if set to limit the search to below the current
   toplevel item
b) select the proper toplevel item to expand to the descendant in the
   following order:
   1) any mounted device or network resource
   2) the user's desktop directory
   3) the user's home directory
   4) the root filesystem