Use only open selection instead of open action with KEY_Up and KEY_Down
If the user navigates the treeview using the up and down keys and a
device is not mounted, then it will be mounted. It is better for user
experience to only mount the device when the user expands the path
with Key_Right, otherwise some items in the treeview might not even
be accessible without accidentally mounting other devices.