Allow opening by middle-click in tree view only for mounted devices and resources
It is complicated to make this working properly for unmounted resources;
The new view will not show the correct location. So let's simply do nothing
when the user clicks on an unmounted resource; the user can click on the
item with the left mouse button to mount it, then middle-click it to open
it in a new tab or window.