Use Alt_L instead of F10 to toggle the menubar (bug #10250)
Other applications like firefox, thunderbird etc. use Alt_L to toggle
a hidden menubar. This seems to be some kind of standard and is also
more intuitive because menu mnemonics are triggered by Alt_L too.

The second part of this commit fixes hiding the menubar (if set to
be invisible).

Of course, the setting still needs to be activated in the view menu or
via Control + M.