Fix loading the trash bin status at startup (bug #9513)
If the gvfsd-trash daemon has not spawned and thunar accesses the trash bin
"trash:///" the first time, then the trash bin does not report the correct
state. This is probably because the daemon needs some time to spawn and get
ready.

As a solution, lets reload the trash file in idle when it is first created,
and then again a bit later after a few seconds have passed. By that time
the gvfs trash daemon should have spawned and the state is properly set.
This is only necessary when first loading the file, not when getting it
from the cache, so we put this into main to be run at startup.