Rework how location bar selection works
ThunarLocationBar now isn't an interface, but an actual widget hosting the
different location bars. Unfortunately, there is still some code left in
ThunarWindow, but this is expected to go away when GtkAction finally dies.

On the plus side, we got rid of GtkUIManager inside the location bars
and also nuked ThunarLocationDialog from orbit.