Use 'folder' icon instead of 'inode-directory'
Actually this is a workaround for icon themes such as Adwaita
that have no 'inode-directory' icon. Previously on gtk2,
gtk_icon_theme_has_icon returned FALSE for 'inode-directory'
while using Adwaita, this is not the case for gtk3, probably
because it falls back to Gnome icon theme, not really desired.