Replace Trash action with Delete as needed (Bug #15352)
We keep the grayed out "Move to Trash" action in directories which are
not writable, but otherwise we show the "Delete" action for files that
cannot be trashed.

Regardless of the "Move to Trash" action's visibility, it should remain
sensitive for keyboard shortcuts to work (i.e.: pressing the Delete key).

Follow-up fix for https://bugzilla.xfce.org/show_bug.cgi?id=15352.