Prevent Gtk-CRITICAL when adding or modifying a user customizable action.
Bug #15970

Adds a check for a NULL value which occurs when adding a new user customizable
action or adding a keyboard shortcut to one which previously did not have one.
This NULL value itself is not an error, since it indicates the abscence of an
existing shortcut, but the lack of this check causes a Gtk-CRITICAL error.