Use XfceFilenameInput
Change the create, rename, and properties dialogs to use the new
XfceFilenameInput widget.
This provies a better user experience when entering filenames. This use
of this widget will
also allow as-you-type rejection of filenames that are too long to be
implemented in the
future, as requested in bugzilla bug #13720.
(Merge request !9)