Add features to support the use of directory specific view settings (#8)
Adds a preference to thunar-preferences to represent the use of directory specific settings,
and adds a checkbox in thunar-preferences-dialog to control it. Adds to thunar-file
and thunarx-file-info to allow the storage of directory specific view settings via gvfs
metadata. Adds a function to thunar-gio-extensions to detect if gvfs metadata support
is present. Does not implement any directory specific settings, just adds features to make
this possible.