Add features to support the use of per directory viewing settings (#8)
Adds a preference to thunar-preferences to represent the use of per directory settings,
and adds a checkbox in thunar-preferences-dialog to control it. Adds to thunar-file
and thunarx-file-info to allow the storage of per directory view settings via gvfs
metadata. Adds a function to thunar-gio-extensions to detect if gvfs metadata support
is present. Does not implement any per directory settings, just adds features to make
this possible.