Remove tray icon and related methods (Issue #495)
"gtk_status_icon_*" is deprecated. It would be possible to replace it
with "StatusNotifier".

Though a tray icon is not really needed. There is still the extra-window
for the copy operation, which should be sufficient.

And the deprecated "gtk_status_icon_new_from_icon_name" seems to cause
scaling issues.

Fixes #495