Fix double-click autoresize on name column (#545)
This commit fixes double-click autoresize (resizing to reasonable fixed
width by double-clicking resizer). This is implemented by abusing
current GTK implementation (See Issue #545 for details), so if auto-resize
for ellipsized text gets fixed, this contribution should be removed.

This commit also unsets manually set fixed-width on directory change.
This is implemented on thunar_details_view_name_resizer(), so if someone
introduces per-file fixed width save, this callback should be modified.

Finally, if the name column is too wide that resizer is not shown in the
screen, it would try to automatically resize one to fit in the screen.

Warning: This commit introduces special THUNAR_COLUMN_MARGIN to prevent
last valid column being autoresized. DO NOT hide this column.