thunar-179: Split thunar-uca-model.c
The first thing I did here is to split out the congested thunar-uca-model.c into
2 other files:

thunar-uca-model-item.c: This entity is used to manage things related to items.
It is also used to write items to files and for updating a custom action item when it changes

thunar-uca-parser.c: This file takes care of all the work of parsing a given uca xml file and feeding
that data back to the model.

Finally, there were further changes made within the model itself from changing the way it saves actions
to how it finds actions from the filesystem