Add a button to install new software to the select application dialog to open file.
This uses the old GNOME PackageKit dbus interface to install
packages according to the mimetype. This interface is implemented
by gnome-software, and now also to xings-packagekit.

Also use dbus to check if there is any installed service that you
can use before adding the corresponding button, therefore it will
not create confusion for other users.