Use move action when dragging files of different users (#196)
Prior this change, if the source file belongs to a user different than
the current one, the default action when dragging that file is copy.
This behavior was introduced long ago in 8bacc1eb, but is not what
users expect.

This commit partly reverts what was introduced by 8bacc1eb, copying when
files are in different filesystems is preserved but now we also check if
the source file is writable (i.e. can be deleted) otherwise move is not
possible.