Use move action when dragging files of different users (#196)
Prior this change, if the source file belongs to a user different than
the current one, the default action when dragging that file is copy.
This behavior was introduced long ago in 8bacc1eb, but it not the
expected one by users.

This commit partly reverts what was introduced by 8bacc1eb, copying when
files are in different filesystems is preserved.