Turn on range checking
This adds a new criteria when checking if an action can be ignored.
First it parses the range string using '-' as a delimiter. If elements
on both sides of it exist, they are converted to integers and compared
to the number of files in the list for the action. Perhaps there is
already code to do this in a more sophistocated way. E.g. in programs
that ask for printer options.

Signed-off-by: Connor Behan <connor.behan@gmail.com>