Add UCA parameter to specify a number of files (Issue #26)
This adds a string to the model and a new entry in the preferences
dialog for setting it. It is designed to store upper and lower limits
for the number of files to which a custom action may be applied. When
checking if an action can be ignored, the code parses the string using
'-' as a delimiter. If elements on both sides of it exist, they are
converted to integers and compared to the number of files in the list
for the action. Perhaps there is already code to do this in a more
sophistocated way. E.g. in programs that ask for printer options.

MR !175