Add Recursive Search (!146)
The recursive search is implemented as a job to not slow down the main thread and maintain responsiveness.
The search job inserts its findings in a list (files_to_add). Every X milliseconds the main thread checks
this list and adds the entirety of its contents to the model of the view where the search takes place.

The search continues in the background if the user changes tabs (this creates a need for the UI to re-enter
search mode when the user changes back to the search tab).

MR !146