Skip files with IO errors during directory scan (Issue #696)
This is needed to avoid the user ending up with an empty directory when that directory contains a corrupted file or a disconnected network drive/mount.

Thunar will display only the accessible files and skip the ones with IO errors.

No error message will be shown to the user.

This new behaviour is similar to Nautilus & Pacman File Manager behaviours in the same situation