WIP: More granular control on weather thumbnails should be requested.
Currently "thunar_standard_view_schedule_thumbnail_idle" is triggered on
each "thunar_standard_view_reload". This will force a thumbnail request
new thumbnails for all files which are visible. (no "lazy flag" used =
"force")

However if no thumbnail is avialabe, it only makes sense to
force-request a new thumbnail when a file actually changed. (Or when the
user hits F5) Wheather the containing folder changed is not relevant for
that decision.

(The current code just disables "lazy" thumbnailing, which actually is
wrong .. planning to use a "force" flag instead of the "lazy" flag for
better understanding)