(.gitignore) ignore thunar.a
MR !268

Since !153 is merged, thunar.a, which is a fake library used for
generating documentation, has been managed by make but not been
covered by .gitignore. This has caused useless yet big file thunar.a
keep appearing in commits that included files with "git add --all".

This commit adds thunar/thunar.a to .gitignore to fix this issue.