Bulk renamer: Rename in multiple iterations (Issue #108)
If file-names which are to be renamed as well are picked as new target-names for other files, it can happen that renaming fails for some of the files.

In order to prevent that, renaming now is done in up to 3 iterations. After the first attempt, the files to rename are sorted in ascending, after 2nd attempt in descending order. After the 3rd attempt, all conflicts should be resolved.

MR !267