preview sidebar: only load the thumbnails if previews wanted.
The images for the sidebar previews should only be loaded or
queued if the user wants those previews visible. Otherwise
it just makes more work that, for some reason, is handled
in the UI thread.

This commit removes an appreciable bit of lag when navigating
around using the cursor keys in folders full of images, even
when the xx-large thumbnails have been generated and are in the
user's cache.

I think it's the gdk_pixbuf_new_from_file() call in this handler
that's actually causing the lag, but for now let's just respect
the preference to keep things speedy for users who don't want the
new feature (people like me...)