New command line option '--separate-window' short '-s'
This command line option enforces opening of a new window
with tabs for each passed location. Additionally a window
opened with this option enabled gets locked. That means
no additional tabs can be added from command line to this
window. Opening new tabs from inside the window or moving tabs
via DnD is still possible. This command line option has
precedence over the similar setting available in the settings
dialog ('open windows as tabs'), so works independent of
the settings dialog setting.

If the settings dialog setting mentioned above is enabled,
the behaviour is as follows:
If unlocked windows are available, the tabs will be added
to the first found unlocked window, else a new window is opened
and tabs are added to this new window.