Dont destroy ThunarFiles on request (#1357)
To just send the "DESTROY" signal on "dispose" and than directly call "finalize" is a bad idea, since e.g. ThunarFolder only removes the ThunarFile from the collection after some delay. Instead, just rely on other objects to release the ThunarFile, and as such the reference count will drop to 0 and finalize will be called.